add: refactor consonants.rs
