+++ src/lingo/mod.rs
@@ -1,1 +1,2 @@
 pub mod past_participle;
+pub mod consonants;
